Fixes issue #363.
Misses the safety check. When you are about to bookmark something for the first time, the request sent to the server contains the stored field by invoking model.get('stored'). Unfortunately, it does not account for the undefined state of the variable and performs operations on it, which results in an unexpected error.

The server replies back with a Bmark object containing the "stored" variable which is used in future communications.